adjust search bar and update styles to not rely on negative margins
