+++ src/styles/sections/icon-info.scss
@@ -5,13 +5,11 @@
   flex-direction: column;
   width: 100%;
   padding: 42px 0;
-  margin-top: -1.5rem;
 
   @include at-media(desktop) {
     flex-direction: row;
     flex-wrap: nowrap;
     justify-content: space-between;
-    margin: 0 -2rem;
   }
 
   .icon-info-section-item {
@@ -34,11 +32,9 @@
       margin-top: 0 !important;
 
       h3 {
-        // width: 100%;
         font-size: size("body", "xl");
         margin: 0;
-        padding: 0;
-        padding-left: units(2);
+        padding: 0 0 0 units(2);
       }
 
       @include at-media(desktop) {
@@ -73,9 +69,19 @@
     }
   }
 
-  .icon-info-section-item:last-of-type {
+  .icon-info-section-item:first-of-type {
+    margin-top: 0;
+
     @include at-media(desktop) {
-      margin-right: -2rem;
+      margin-left: 0;
+    }
+  }
+
+  .icon-info-section-item:last-of-type {
+    margin-bottom: 0;
+
+    @include at-media(desktop) {
+      margin-right: 0;
     }
   }
